+++ features/login/components/loginForm.tsx
@@ -1,26 +1,20 @@
+"use client";
+
 import { Button } from "@/components/button";
 import BaseInput from "@/components/input/baseInput";
 import ErrorMsg from "@/components/text/errorMsg";
-import { UserInfo } from "@/types/user";
+import useAuth from "@/hooks/useAuth";
 
-interface SignFormProps {
-  userInfo: UserInfo;
-  errorMsg: string;
-  handleUserSignUp: (e: React.FormEvent<HTMLFormElement>) => void;
-  handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
+const LoginForm = () => {
+  const { handleUserLogin, errorMsg, userInfo, setUserInfo } = useAuth();
 
-const SignForm = ({ userInfo, errorMsg, handleUserSignUp, handleChange }: SignFormProps) => {
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    setUserInfo({ ...userInfo, [name]: value });
+  };
+
   return (
-    <form onSubmit={handleUserSignUp} className="flex w-full flex-col gap-4">
-      <BaseInput
-        label="이름"
-        placeholder="이름을 입력해주세요."
-        name="name"
-        value={userInfo.name}
-        onChange={handleChange}
-        isRequired
-      />
+    <form onSubmit={handleUserLogin} className="flex w-full flex-col gap-4">
       <BaseInput
         label="이메일"
         placeholder="이메일을 입력해주세요."
@@ -38,15 +32,13 @@ const SignForm = ({ userInfo, errorMsg, handleUserSignUp, handleChange }: SignFo
         value={userInfo.password}
         onChange={handleChange}
         isRequired
-        minLength={8}
-        maxLength={16}
       />
       {errorMsg && <ErrorMsg errorMsg={errorMsg} />}
       <Button variant="secondary" size="medium" type="submit">
-        회원가입
+        로그인
       </Button>
     </form>
   );
 };
 
-export default SignForm;
+export default LoginForm;
