+++ features/signUp/index.tsx
@@ -1,35 +1,20 @@
-"use client";
-
 import ButttonContain from "@/components/button/butttonContain";
 import ScreenContainer from "@/components/container/screenContainer";
 import ActionLinkMsg from "@/components/text/actionLinkMsg";
 import CircleTitle from "@/components/title/circleTitle";
 import Title from "@/components/title/title";
 import { routes } from "@/constants/path";
-import useAuth from "@/hooks/useAuth";
 
-import SignForm from "./signForm";
+import SignForm from "./components/signForm";
 
 const SignUpScreen = () => {
-  const { handleUserSignUp, errorMsg, userInfo, setUserInfo } = useAuth();
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setUserInfo({ ...userInfo, [name]: value });
-  };
-
   return (
     <ScreenContainer>
       <CircleTitle>
         <Title size="text-xl">정보를 입력해주세요!</Title>
       </CircleTitle>
 
-      <SignForm
-        handleUserSignUp={handleUserSignUp}
-        userInfo={userInfo}
-        handleChange={handleChange}
-        errorMsg={errorMsg}
-      />
+      <SignForm />
 
       <ButttonContain isFixed isColumn>
         <ActionLinkMsg
